requirements.txt
boto3
kafka-python
smart_open
